+++ m2/benchmarks/binaryConverter/binaryConverter.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 #include<stdlib.h>
-long _wait(long EV_waitTime)
+int _wait(int EV_waitTime)
 {
 while ((EV_waitTime>0))
 {
@@ -8,9 +8,9 @@ EV_waitTime = (EV_waitTime-1);
 }
 return 0;
 }
-long _power(long EV_base,long EV_exponent)
+int _power(int EV_base,int EV_exponent)
 {
-long EV_product;
+int EV_product;
 EV_product = 1;
 while ((EV_exponent>0))
 {
@@ -19,11 +19,11 @@ EV_exponent = (EV_exponent-1);
 }
 return EV_product;
 }
-long _recursiveDecimalSum(long EV_binaryNum,long EV_decimalSum,long EV_recursiveDepth)
+int _recursiveDecimalSum(int EV_binaryNum,int EV_decimalSum,int EV_recursiveDepth)
 {
-long EV_tempNum;
-long EV_base;
-long EV_remainder;
+int EV_tempNum;
+int EV_base;
+int EV_remainder;
 if ((EV_binaryNum>0))
 {
 EV_base = 2;
@@ -38,19 +38,19 @@ return _recursiveDecimalSum((EV_binaryNum/10), EV_decimalSum, (EV_recursiveDepth
 }
 return EV_decimalSum;
 }
-long _convertToDecimal(long EV_binaryNum)
+int _convertToDecimal(int EV_binaryNum)
 {
-long EV_recursiveDepth;
-long EV_decimalSum;
+int EV_recursiveDepth;
+int EV_decimalSum;
 EV_recursiveDepth = 0;
 EV_decimalSum = 0;
 return _recursiveDecimalSum(EV_binaryNum, EV_decimalSum, EV_recursiveDepth);
 }
-long _main()
+int _main()
 {
-long EV_number;
-long EV_waitTime;
-scanf("%ld", &EV_number);
+int EV_number;
+int EV_waitTime;
+scanf("%d", &EV_number);
 EV_number = _convertToDecimal(EV_number);
 EV_waitTime = (EV_number*EV_number);
 while ((EV_waitTime>0))
@@ -58,7 +58,7 @@ while ((EV_waitTime>0))
 _wait(EV_waitTime);
 EV_waitTime = (EV_waitTime-1);
 }
-printf("%ld\n", (long)EV_number);
+printf("%d\n", EV_number);
 return 0;
 }
 int main(void)
