+++ m2/benchmarks/killerBubbles/killerBubbles.c
@@ -2,19 +2,19 @@
 #include<stdlib.h>
 struct EV_Node
 {
-long EV_val;
+int EV_val;
 struct EV_Node * EV_prev;
 struct EV_Node * EV_next;
 };
-long EV_swapped;
-long _compare(struct EV_Node * EV_a,struct EV_Node * EV_b)
+int EV_swapped;
+int _compare(struct EV_Node * EV_a,struct EV_Node * EV_b)
 {
 return (EV_a->EV_val-EV_b->EV_val);
 }
 void _deathSort(struct EV_Node * EV_head)
 {
-long EV_swapped;
-long EV_swap;
+int EV_swapped;
+int EV_swap;
 struct EV_Node * EV_currNode;
 EV_swapped = 1;
 while ((EV_swapped==1))
@@ -39,28 +39,28 @@ void _printEVILList(struct EV_Node * EV_head)
 struct EV_Node * EV_currNode;
 struct EV_Node * EV_toFree;
 EV_currNode = EV_head->EV_next;
-printf("%ld\n", (long)EV_head->EV_val);
+printf("%d\n", EV_head->EV_val);
 free(EV_head);
 while ((EV_currNode!=EV_head))
 {
 EV_toFree = EV_currNode;
-printf("%ld\n", (long)EV_currNode->EV_val);
+printf("%d\n", EV_currNode->EV_val);
 EV_currNode = EV_currNode->EV_next;
 free(EV_toFree);
 }
 }
-long _main()
+int _main()
 {
-long EV_numNodes;
-long EV_counter;
+int EV_numNodes;
+int EV_counter;
 struct EV_Node * EV_currNode;
 struct EV_Node * EV_head;
 struct EV_Node * EV_previous;
 EV_swapped = 666;
-scanf("%ld", &EV_numNodes);
+scanf("%d", &EV_numNodes);
 if ((EV_numNodes<=0))
 {
-printf("%ld\n", (long)(-1));
+printf("%d\n", (-1));
 return (-1);
 }
 EV_numNodes = (EV_numNodes*1000);
