+++ m2/benchmarks/mixed/mixed.c
@@ -2,16 +2,16 @@
 #include<stdlib.h>
 struct EV_simple
 {
-long EV_one;
+int EV_one;
 };
 struct EV_foo
 {
-long EV_bar;
-long EV_cool;
+int EV_bar;
+int EV_cool;
 struct EV_simple * EV_simp;
 };
 struct EV_foo * EV_globalfoo;
-void _tailrecursive(long EV_num)
+void _tailrecursive(int EV_num)
 {
 struct EV_foo * EV_unused;
 if ((EV_num<=0))
@@ -21,15 +21,15 @@ return ;
 EV_unused = (struct EV_foo*)malloc(sizeof(struct EV_foo));
 _tailrecursive((EV_num-1));
 }
-long _add(long EV_x,long EV_y)
+int _add(int EV_x,int EV_y)
 {
 return (EV_x+EV_y);
 }
-void _domath(long EV_num)
+void _domath(int EV_num)
 {
 struct EV_foo * EV_math1;
 struct EV_foo * EV_math2;
-long EV_tmp;
+int EV_tmp;
 EV_math1 = (struct EV_foo*)malloc(sizeof(struct EV_foo));
 EV_math1->EV_simp = (struct EV_simple*)malloc(sizeof(struct EV_simple));
 EV_math2 = (struct EV_foo*)malloc(sizeof(struct EV_foo));
@@ -49,7 +49,7 @@ EV_num = (EV_num-1);
 free(EV_math1);
 free(EV_math2);
 }
-void _objinstantiation(long EV_num)
+void _objinstantiation(int EV_num)
 {
 struct EV_foo * EV_tmp;
 while ((EV_num>0))
@@ -59,7 +59,7 @@ free(EV_tmp);
 EV_num = (EV_num-1);
 }
 }
-long _ackermann(long EV_m,long EV_n)
+int _ackermann(int EV_m,int EV_n)
 {
 if ((EV_m==0))
 {
@@ -74,25 +74,25 @@ else
 return _ackermann((EV_m-1), _ackermann(EV_m, (EV_n-1)));
 }
 }
-long _main()
+int _main()
 {
-long EV_a;
-long EV_b;
-long EV_c;
-long EV_d;
-long EV_e;
-scanf("%ld", &EV_a);
-scanf("%ld", &EV_b);
-scanf("%ld", &EV_c);
-scanf("%ld", &EV_d);
-scanf("%ld", &EV_e);
+int EV_a;
+int EV_b;
+int EV_c;
+int EV_d;
+int EV_e;
+scanf("%d", &EV_a);
+scanf("%d", &EV_b);
+scanf("%d", &EV_c);
+scanf("%d", &EV_d);
+scanf("%d", &EV_e);
 _tailrecursive(EV_a);
-printf("%ld\n", (long)EV_a);
+printf("%d\n", EV_a);
 _domath(EV_b);
-printf("%ld\n", (long)EV_b);
+printf("%d\n", EV_b);
 _objinstantiation(EV_c);
-printf("%ld\n", (long)EV_c);
-printf("%ld\n", (long)_ackermann(EV_d, EV_e));
+printf("%d\n", EV_c);
+printf("%d\n", _ackermann(EV_d, EV_e));
 return 0;
 }
 int main(void)
