+++ m2/benchmarks/primes/primes.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 #include<stdlib.h>
-long _isqrt(long EV_a)
+int _isqrt(int EV_a)
 {
-long EV_square;
-long EV_delta;
+int EV_square;
+int EV_delta;
 EV_square = 1;
 EV_delta = 3;
 while ((EV_square<=EV_a))
@@ -13,11 +13,11 @@ EV_delta = (EV_delta+2);
 }
 return ((EV_delta/2)-1);
 }
-long _prime(long EV_a)
+int _prime(int EV_a)
 {
-long EV_max;
-long EV_divisor;
-long EV_remainder;
+int EV_max;
+int EV_divisor;
+int EV_remainder;
 if ((EV_a<2))
 {
 return 0;
@@ -38,17 +38,17 @@ EV_divisor = (EV_divisor+1);
 return 1;
 }
 }
-long _main()
+int _main()
 {
-long EV_limit;
-long EV_a;
-scanf("%ld", &EV_limit);
+int EV_limit;
+int EV_a;
+scanf("%d", &EV_limit);
 EV_a = 0;
 while ((EV_a<=EV_limit))
 {
 if (_prime(EV_a))
 {
-printf("%ld\n", (long)EV_a);
+printf("%d\n", EV_a);
 }
 EV_a = (EV_a+1);
 }
