+++ m2/benchmarks/programBreaker/programBreaker.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 #include<stdlib.h>
-long EV_GLOBAL;
-long EV_count;
-long _fun2(long EV_x,long EV_y)
+int EV_GLOBAL;
+int EV_count;
+int _fun2(int EV_x,int EV_y)
 {
 if ((EV_x==0))
 {
@@ -13,9 +13,9 @@ else
 return _fun2((EV_x-1), EV_y);
 }
 }
-long _fun1(long EV_x,long EV_y,long EV_z)
+int _fun1(int EV_x,int EV_y,int EV_z)
 {
-long EV_retVal;
+int EV_retVal;
 EV_retVal = ((((5+6)-(EV_x*2))+(4/EV_y))+EV_z);
 if ((EV_retVal>EV_y))
 {
@@ -30,14 +30,14 @@ return _fun2(EV_retVal, EV_y);
 }
 return EV_retVal;
 }
-long _main()
+int _main()
 {
-long EV_i;
+int EV_i;
 EV_i = 0;
-scanf("%ld", &EV_i);
+scanf("%d", &EV_i);
 while ((EV_i<10000))
 {
-printf("%ld\n", (long)_fun1(3, EV_i, 5));
+printf("%d\n", _fun1(3, EV_i, 5));
 EV_i = (EV_i+1);
 }
 return 0;
